Fin paso 41

+++ database/migrations/2021_11_03_151203_create_user_profiles_table.php
@@ -14,13 +14,13 @@ class CreateUserProfilesTable extends Migration
     public function up()
     {
         Schema::create('user_profiles', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
 
             $table->string('bio', 1000);
             $table->string('twitter')->nullable();
-            $table->unsignedInteger('profession_id')->nullable();
+            $table->unsignedBigInteger('profession_id')->nullable();
             $table->foreign('profession_id')->references('id')->on('professions');
-            $table->unsignedInteger('user_id')->unique();
+            $table->unsignedBigInteger('user_id')->unique();
             $table->foreign('user_id')->references('id')->on('users')
                 ->onDelete('CASCADE');
