+++ database/migrations/2021_11_23_153415_add_team_id_to_users.php
@@ -14,7 +14,7 @@ class AddTeamIdToUsers extends Migration
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->unsignedInteger('team_id')->nullable()->after('id');
+            $table->unsignedBigInteger('team_id')->nullable()->after('id');
             $table->foreign('team_id')->references('id')->on('teams');
         });
     }
